requirements.txt
gradio
pandas
faiss-cpu
python-dotenv
google-generativeai
langchain
langchain-community
sentence-transformers
